Change default user-agent string to be compatible with newer Cisco servers
See https://gitlab.com/openconnect/openconnect/-/issues/665 for a summary of
this issue.

This implements the simplest reasonable solution to the problem: Just Change
The Default™ UA string.

Short summary: Cisco did something stupidly backwards-incompatible in their
authentication flow.  It's hard to tell if it was due to incompetence or due
to malice towards unofficial clients
(https://gitlab.com/openconnect/openconnect/-/issues/635#note_1451782874)
but it doesn't really matter.

If merged, this should fix #544, #593, #602, #618, #635, #657, #662,
and #665.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>