csd-wrapper: remove trailing part from URL
The Cisco ASA server I use got some update.
It suddenly stopped providing the reply:
	<host-scan-base-uri>/CACHE</host-scan-base-uri>
and went back to the legacy reply:
	<csdLinux
		stuburl="/CACHE/sdesktop/install/binaries/sfinst"
		starturl="/CACHE/sdesktop/install/result.htm"
		waiturl="/+CSCOE+/sdesktop/wait.html"
	/>

The csd-wrapper now receives in the command line
	-url https://server/CACHE/sdesktop/install/result.htm
instead of
	-url https://server/CACHE
and, obviously, the concatenation of $URL within the script makes
impossible to download the manifest.

Strip away the trailing "/sdesktop/*" from $URL.
This should work even in case of custom path in the ASA server.

Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>