Fix Cisco DTLS MTU detection
New ASA firmware apparently doesn't send back the same payload in a DPD
response, so we can't match on it. But it does thankfully still return
responses of the same length, it seems. And the matching isn't important;
if *any* packet of the given size gets through, that's good enough for
the MTU detection. So just match the length.

And once we start looking at it, it looks like the whole thing was hosed
and would never even retry the same packet size as it was supposed to.
Fix it all.

Fixes: #823

Signed-off-by: David Woodhouse <dwmw2@infradead.org>